dynamodb table tests and create endpoint tests

+++ tests/test_comments/test_entities.py
@@ -20,4 +20,4 @@ def test_serialize(dummy_comment: Comment) -> None:
     # then
     assert isinstance(comment["id"], str)
     assert isinstance(comment["post_id"], str)
-    assert isinstance(comment["created_at"], float)
+    assert isinstance(comment["created_at"], int)
